[refactor] 인기대출도서 API 관련 요청, 응답 DTO 통합
