+++ src/main/java/com/book/backend/domain/genre/controller/GenreController.java
@@ -1,14 +1,9 @@
 package com.book.backend.domain.genre.controller;
 
-import com.book.backend.domain.book.service.BookRequestValidate;
 import com.book.backend.domain.genre.service.GenreRequestValidate;
 import com.book.backend.domain.genre.service.GenreService;
-import com.book.backend.domain.openapi.dto.request.genre.NewTrendRequestDto;
-import com.book.backend.domain.openapi.dto.request.genre.PeriodTrendRequestDto;
-import com.book.backend.domain.openapi.dto.request.genre.RandomRequestDto;
-import com.book.backend.domain.openapi.dto.response.genre.NewTrendResponseDto;
-import com.book.backend.domain.openapi.dto.response.genre.PeriodTrendResponseDto;
-import com.book.backend.domain.openapi.dto.response.genre.RandomResponseDto;
+import com.book.backend.domain.openapi.dto.request.LoanTrendRequestDto;
+import com.book.backend.domain.openapi.dto.response.LoanTrendResponseDto;
 import com.book.backend.global.ResponseTemplate;
 import com.book.backend.global.log.RequestLogger;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +13,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.LinkedList;
-import java.util.Random;
 
 @RestController
 @RequestMapping("/api/genre")
@@ -38,8 +32,8 @@ public class GenreController {
         RequestLogger.param(new String[]{"kdcNum"}, subKdc);
         genreRequestValidate.isValidSubKdc(subKdc);
 
-        PeriodTrendRequestDto requestDto = PeriodTrendRequestDto.builder().dtl_kdc(subKdc).build();
-        LinkedList<PeriodTrendResponseDto> response = genreService.periodTrend(requestDto, 7, maxSize);
+        LoanTrendRequestDto requestDto = LoanTrendRequestDto.builder().dtl_kdc(subKdc).build();
+        LinkedList<LoanTrendResponseDto> response = genreService.periodTrend(requestDto, 7, maxSize);
 
         return responseTemplate.success(response, HttpStatus.OK);
     }
@@ -53,8 +47,8 @@ public class GenreController {
         RequestLogger.param(new String[]{"kdcNum"}, subKdc);
         genreRequestValidate.isValidSubKdc(subKdc);
 
-        PeriodTrendRequestDto requestDto = PeriodTrendRequestDto.builder().dtl_kdc(subKdc).build();
-        LinkedList<PeriodTrendResponseDto> response = genreService.periodTrend(requestDto, 30, maxSize);
+        LoanTrendRequestDto requestDto = LoanTrendRequestDto.builder().dtl_kdc(subKdc).build();
+        LinkedList<LoanTrendResponseDto> response = genreService.periodTrend(requestDto, 30, maxSize);
 
         return responseTemplate.success(response, HttpStatus.OK);
     }
@@ -68,8 +62,8 @@ public class GenreController {
         RequestLogger.param(new String[]{"kdcNum"}, subKdc);
         genreRequestValidate.isValidSubKdc(subKdc);
 
-        RandomRequestDto requestDto = RandomRequestDto.builder().dtl_kdc(subKdc).build();
-        LinkedList<RandomResponseDto> response = genreService.random(requestDto, maxSize);
+        LoanTrendRequestDto requestDto = LoanTrendRequestDto.builder().dtl_kdc(subKdc).build();
+        LinkedList<LoanTrendResponseDto> response = genreService.random(requestDto, maxSize);
 
         return responseTemplate.success(response, HttpStatus.OK);
     }
@@ -83,8 +77,8 @@ public class GenreController {
         RequestLogger.param(new String[]{"kdcNum"}, subKdc);
         genreRequestValidate.isValidSubKdc(subKdc);
 
-        NewTrendRequestDto requestDto = NewTrendRequestDto.builder().dtl_kdc(subKdc).build();
-        LinkedList<NewTrendResponseDto> response = genreService.newTrend(requestDto, maxSize);
+        LoanTrendRequestDto requestDto = LoanTrendRequestDto.builder().dtl_kdc(subKdc).build();
+        LinkedList<LoanTrendResponseDto> response = genreService.newTrend(requestDto, maxSize);
 
         return responseTemplate.success(response, HttpStatus.OK);
     }
