--- src/main/java/com/book/backend/domain/openapi/dto/request/genre/RandomRequestDto.java
@@ -1,16 +1,0 @@
-package com.book.backend.domain.openapi.dto.request.genre;
-
-import com.book.backend.domain.openapi.dto.request.OpenAPIRequestInterface;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class RandomRequestDto implements OpenAPIRequestInterface {
-    @NotBlank(message = "subKdc는 필수 입력값입니다.")
-    private String dtl_kdc;
-    private Integer pageSize;  // 페이지 사이즈
-}
